Modified how threads is handled
